+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - `RandomizeWeapons` now only picks a random weapon that you have unlocked instead of any weapon. It also only picks a random aspect that you have unlocked, instead of any aspect.
+- `RandomizeKeepsakes` now only picks unlocked keepsakes instead of any keepsake.
 
 ## [1.4.0] - 2024-07-15